pyproject.toml
[project]
name = "iis"
version = "0.1.0"
description = "Projekt za Predem IIS"
authors = [
    {name = "Blazhe Manev",email = "blaze_manev@yahoo.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy (>=2.2.4,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "lxml (>=5.3.2,<6.0.0)",
    "pyaml (>=25.1.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
